requirements.txt
python3~=3.11.3
requests~=2.30.0
python-dotenv~=1.0.0